Move runner class to a new file
